refactor: clear all 143 remaining compiler warnings
No behaviour change. Every edit here is a construct the Kotlin compiler
proved could not do anything.

The bulk (157 edits across 28 files) is redundant `?.` and `!!` on
receivers that were already non-null — mostly the `optString(...)?.` and
`!!`-on-Int idioms. These were applied by script from the compiler's own
file:line:column output, after asserting that every one of the 137
reported columns pointed at exactly the `?.` or `!!` it claimed to; zero
mismatches, so the transform was targeted rather than a regex guess. It
took two passes: removing one operator in a chain exposes the next.

That in turn exposed a second layer of 10 findings the safe calls had
been masking, each of which needed a judgement call rather than a
deletion:

- Dead elvis fallbacks on non-null receivers (AppViewModel room-list
  hashing, ConversationsApi shortcut label, SyncRoomsCoordinator invites
  and roomsWithEvents, SyncIngestor is_direct, MediaUploadUtils and
  MediaFunctions stream handling, NarratorFunctions kick narration).
- NarratorFunctions carried a whole `?: run { }` block captioned
  "Fallback if leftUserId is somehow null", three lines below a comment
  asserting leftUserId is guaranteed non-null in that branch. It could
  never run.
- UrlPreviewComposition's `if (body == null)` guard, dead once
  response.body.string() is known non-null. An empty body still falls to
  the existing catch, as before.

The remainder: redundant .toFloat() on a Float (MainActivity, UserInfo),
redundant `as AnnotatedString` on subSequence (html.kt), a redundant
`else` on an exhaustive sealed `when` (BubbleTimelineScreen), and an
explicit @OptIn for maps-compose's MapEffect, which is incubating but is
the only route to setOnPoiClickListener.

That last one also moved a detekt-baseline entry: the baseline keys
findings by a signature that includes the annotation list, so extending
@OptIn resurfaced a pre-existing ModifierMissing finding on the same
unchanged function. Re-pointed rather than re-suppressed.

Verified with a forced full recompile (an up-to-date task emits no
warnings at all, which reads as a false clean), then detekt, unit tests,
and clean assembleBaseDebug: 0 warnings, 0 errors.

+++ app/src/main/java/net/vrkknn/andromuks/MainActivity.kt
@@ -112,7 +112,7 @@ private fun rememberMorphingStartupAvatarMaskModifier(): Modifier {
     val scaled = morphCycle * segmentCount
     val fromIndex = kotlin.math.floor(scaled).toInt().mod(segmentCount)
     val toIndex = (fromIndex + 1).mod(segmentCount)
-    val localProgress = scaled - kotlin.math.floor(scaled).toFloat()
+    val localProgress = scaled - kotlin.math.floor(scaled)
     val morph = remember(fromIndex, toIndex) { Morph(shapes[fromIndex], shapes[toIndex]) }
     val rawPath = remember(morph, localProgress) {
         morph.toPath(progress = localProgress).asComposePath()
